correction du processus de lancement et avant nettoyage

+++ skini.js
@@ -14,47 +14,31 @@ function startSkini() {
 
   // Nettoyage du log précédents
   try {
-    fs.unlinkSync("./skinilog.json")
-    //file removed
-  } catch (err) {
-    console.error("WARN: Pas de log précédents:", err)
+    if (fs.existsSync("./skinilog.json")) {
+      try {
+        fs.unlinkSync("./skinilog.json")
+      } catch (err) {
+        console.error("WARN: Pas de log précédents:", err)
+      }
+    }
+  } catch(err) {
+    console.error(err)
   }
 
   // Websocket dans le Serveur
   var ws = require('./serveur/websocketServer');
   var oscReceiveDAW = require("./serveur/midimix.js");
-  var par = require('./serveur/skiniParametres');
+  //var par = require('./serveur/skiniParametres'); //!!!!
 
-  // To share the parameters file at launch. ws dispaches to oscMidi,
-  // DAW and groupesClientsSons
-  ws.setParameters(par, oscReceiveDAW);
+  // Load the necessary modules in the websocket server at launch
+  ws.setParameters(oscReceiveDAW);
 
   // Share ws server in midimix
   oscReceiveDAW.setWebSocketServer(ws);
 
   // AFFICHAGE DU CONTEXTE =================================================
 
-  function displayContext(param, ipConfig) {
-    // console.log("\n=========================================================");
-    // //if(param.scenesON) console.log("Skini avec les scènes\n");
-    // if (param.avecMusicien) {
-    //   console.log("Décalage des FIFOs de ", param.decalageFIFOavecMusicien, " pulsations");
-    // }
-    // console.log("Config des patterns: ", param.configClips);
-    // console.log("Fichiers sons dans : ", param.soundFilesPath1);
-    // if (param.avecMusicien) console.log("Fichiers partitions dans : ", param.patternScorePath1);
-
-    // console.log("Synchro. Midi: ", param.synchoOnMidiClock);
-    // console.log("Réaction au moment où le pattern est joué (par défaut à la sélection):", param.reactOnPlay);
-    // console.log("Algorithme sur FIFO: ", param.algoGestionFifo);
-    // console.log("Nombe de groupes clients: ", param.nbeDeGroupesClients);
-
-    // console.log("----------- SIMULATEUR ----------------------------------");
-    // console.log("Simulation dans un groupe dédié: ", param.simulatorInAseperateGroup);
-    // console.log("Tempo max simulateur en ms:", param.tempoMax);
-    // console.log("Tempo min simulateur en ms:", param.tempoMin);
-    // console.log("Limite de durée d'attente en pulsations :", param.limiteDureeAttente);
-
+  function displayContext(ipConfig) {
     console.log("----------- RESEAU --------------------------------------");
     console.log("Port OUT OSC pour MIDI:", ipConfig.OutPortOSCMIDItoDAW, "IP:", ipConfig.remoteIPAddressSound); // DAW ou Processing
     console.log("Port IN  OSC pour MIDI:", ipConfig.InPortOSCMIDIfromDAW, "IP:", ipConfig.remoteIPAddressSound); // DAW ou Processing
@@ -74,14 +58,10 @@ function startSkini() {
     }
 
     console.log("----------- MIDI --------------------------------------");
-    if (par.directMidiON) {
-      console.log("Use MIDI direct to control DAW");
-    } else {
-      console.log("Use OSC to control DAW");
-    }
     for (var i = 0; i < midiConfig.length; i++) {
       console.log("Midi" + midiConfig[i].type + ", usage:" + midiConfig[i].spec + ", bus: " + midiConfig[i].name + ", " + midiConfig[i].comment);
     }
+
     console.log("=========================================================");
   }
 
@@ -123,7 +103,7 @@ function startSkini() {
     console.log(`INFO: Skini listening at http://${addressServer}:${port}`);
   });
 
-  displayContext(par, ipConfig);
+  displayContext(ipConfig);
 }
 
 startSkini();
